requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
itsdangerous==2.2.0
sqlalchemy==2.0.34
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
cryptography==43.0.1
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
